+++ AsteroidPlayerCollisionResponseSystem/src/main/java/io/asteroidsfx/asteroidplayercollisionresponsesystem/AsteroidPlayerCollisionResponseSystem.java
@@ -31,7 +31,7 @@ public class AsteroidPlayerCollisionResponseSystem extends System {
     }
 
     @Override
-    public void tick(float dt, List<Entity> entities) {
+    public void tick(double dt, List<Entity> entities) {
 
     }
 }
