+++ InputComponent/src/main/java/io/asteroidsfx/inputcomponent/InputAction.java
@@ -4,5 +4,5 @@ import io.asteroidsfx.common.Entity;
 
 @FunctionalInterface
 public interface InputAction{
-    void execute(Entity entity, float dt);
+    void execute(Entity entity, double dt);
 }
