+++ InputSystem/src/main/java/io/asteroidsfx/inputsystem/InputSystem.java
@@ -23,7 +23,7 @@ public class InputSystem extends System{
     }
 
     @Override
-    public void tick(float dt, List<Entity> entities) {
+    public void tick(double dt, List<Entity> entities) {
         for (Entity entity : entities){
             InputComponent inputComponent = entity.getComponent(InputComponent.class);
